+++ src/main/java/ru/practicum/shareit/user/storage/InMemoryUserStorage.java
@@ -10,10 +10,10 @@ import java.util.Map;
 @Component
 public class InMemoryUserStorage implements UserStorage {
 
-    private int uniqueId = 0;
-    public final Map<Integer, User> users = new HashMap<>();
+    private Long uniqueId = Long.valueOf(0);
+    public final Map<Long, User> users = new HashMap<>();
 
-    private int getUniqueId() {
+    private Long getUniqueId() {
         uniqueId++;
         return uniqueId;
     }
